returning promise with wallets

+++ database/database.sql
@@ -1,7 +1,7 @@
 CREATE TABLE wallets (
     id SERIAL PRIMARY KEY,
     name VARCHAR,
-    amount INTEGER,
+    amount FLOAT,
     icon VARCHAR
 );
 
@@ -14,7 +14,7 @@ CREATE TABLE expense_type (
 CREATE TABLE expenses (
     id SERIAL PRIMARY KEY,
     expense_type_id INTEGER REFERENCES expense_type(id),
-    amount INTEGER,
+    amount FLOAT,
     date DATE
 );
 
@@ -27,21 +27,21 @@ CREATE TABLE add_type (
 CREATE TABLE additives (
     id SERIAL PRIMARY KEY,
     add_type_id INTEGER REFERENCES add_type(id),
-    amount INTEGER,
+    amount FLOAT,
     date DATE
 );
 
 CREATE TABLE debt (
     id SERIAL PRIMARY KEY,
     creditor VARCHAR,
-    amount INTEGER,
+    amount FLOAT,
     due_date DATE
 );
 
 CREATE TABLE savings (
     id SERIAL PRIMARY KEY,
     title VARCHAR,
-    amount INTEGER,
+    amount FLOAT,
     last_updated DATE,
     percentage DECIMAL
 );
